Voice2TextBot/requirements.txt
python-telegram-bot[rate-limiter]
transformers
torch
torchaudio
librosa
soundfile
scipy
numpy
